% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{add_provenance_cols}
\alias{add_provenance_cols}
\title{Add provenance columns}
\usage{
add_provenance_cols(data, source_name = NA_character_, rule_id = NA_character_)
}
\arguments{
\item{data}{Tibble.}

\item{source_name}{Character.}

\item{rule_id}{Character.}
}
\value{
Tibble.
}
\description{
Add provenance columns
}
