% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{apply_domain_rules}
\alias{apply_domain_rules}
\title{Execute ordered rules on a domain}
\usage{
apply_domain_rules(data, domain_rules, context, stop_on_error = TRUE)
}
\arguments{
\item{data}{Tibble.}

\item{domain_rules}{List.}

\item{context}{\code{build_context} or list.}

\item{stop_on_error}{Logical. Default \code{TRUE}.}
}
\value{
Named list.
}
\description{
Execute ordered rules on a domain
}
