% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g3_derive_dates.R
\name{apply_imputation_policy}
\alias{apply_imputation_policy}
\title{Apply imputation policy}
\usage{
apply_imputation_policy(parsed, policy = list())
}
\arguments{
\item{parsed}{Tibble from parse_partial_date.}

\item{policy}{List.}
}
\value{
Tibble.
}
\description{
Apply imputation policy
}
