% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_data_access.R
\name{apply_missing_conventions}
\alias{apply_missing_conventions}
\title{Apply missing value conventions}
\usage{
apply_missing_conventions(
  data,
  config = NULL,
  blank_to_na = TRUE,
  unk_tokens = c("UNK", "UNKNOWN", "NK", "N/A", "."),
  unk_to_na = FALSE
)
}
\arguments{
\item{data}{Tibble.}

\item{config}{\code{sdtm_config} or \code{NULL}.}

\item{blank_to_na}{Logical. Default \code{TRUE}.}

\item{unk_tokens}{Character vector.}

\item{unk_to_na}{Logical. Default \code{FALSE}.}
}
\value{
Tibble with standardized missings.
}
\description{
Apply missing value conventions
}
