% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g2_derive_ct.R
\name{assign_ct}
\alias{assign_ct}
\title{Assign controlled terminology coded values}
\usage{
assign_ct(
  data,
  target_var,
  source_var,
  codelist_id,
  ct_lib,
  case_sensitive = FALSE,
  unknown_policy = "warn_and_keep",
  custom_map = NULL
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{codelist_id}{Character.}

\item{ct_lib}{Tibble.}

\item{case_sensitive}{Logical. Default \code{FALSE}.}

\item{unknown_policy}{Character. Default \code{"warn_and_keep"}.}

\item{custom_map}{Named character vector or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Assign controlled terminology coded values
}
