% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{bind_sources}
\alias{bind_sources}
\title{Stack multiple source datasets}
\usage{
bind_sources(..., source_names = NULL, fill_missing = TRUE)
}
\arguments{
\item{...}{Tibbles or a named list.}

\item{source_names}{Character vector.}

\item{fill_missing}{Logical. Default \code{TRUE}.}
}
\value{
Tibble.
}
\description{
Stack multiple source datasets
}
