% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_dependency.R
\name{build_dependency_graph}
\alias{build_dependency_graph}
\title{Build a variable dependency graph for a domain}
\usage{
build_dependency_graph(
  rule_set,
  domain,
  allow_cycles = FALSE,
  late_bind_vars = character()
)
}
\arguments{
\item{rule_set}{\code{rule_set} object.}

\item{domain}{Character.}

\item{allow_cycles}{Logical. Default \code{FALSE}.}

\item{late_bind_vars}{Character vector. Default \code{character()}.}
}
\value{
Named list with \code{graph}, \code{order}, \code{cycles}, \code{phases}.
}
\description{
Build a variable dependency graph for a domain
}
