% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{build_dm_plugin}
\alias{build_dm_plugin}
\title{DM domain plugin}
\usage{
build_dm_plugin(target_meta, source_meta, raw_data, config, rule_set)
}
\arguments{
\item{target_meta}{Tibble.}

\item{source_meta}{Tibble.}

\item{raw_data}{Named list.}

\item{config}{\code{sdtm_config}.}

\item{rule_set}{\code{rule_set}.}
}
\value{
\code{build_result} for DM.
}
\description{
DM domain plugin
}
