% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{build_domain_from_sources}
\alias{build_domain_from_sources}
\title{Build domain from multiple sources}
\usage{
build_domain_from_sources(domain, rule_set, raw_data, source_meta, config)
}
\arguments{
\item{domain}{Character.}

\item{rule_set}{\code{rule_set}.}

\item{raw_data}{Named list of tibbles.}

\item{source_meta}{Tibble.}

\item{config}{\code{sdtm_config}.}
}
\value{
Tibble.
}
\description{
Build domain from multiple sources
}
