% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{build_relrec}
\alias{build_relrec}
\title{Build RELREC dataset}
\usage{
build_relrec(relationship_specs, domain_data, config)
}
\arguments{
\item{relationship_specs}{List.}

\item{domain_data}{Named list.}

\item{config}{\code{sdtm_config}.}
}
\value{
Tibble or \code{NULL}.
}
\description{
Build RELREC dataset
}
