% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{build_supp}
\alias{build_supp}
\title{Build SUPP-- dataset}
\usage{
build_supp(
  domain_data,
  domain,
  target_meta,
  vars_to_supp = NULL,
  idvar = NULL,
  qnam_rules = NULL,
  qlabel_source = "metadata",
  qorig_source = "metadata",
  qeval = NA_character_
)
}
\arguments{
\item{domain_data}{Tibble.}

\item{domain}{Character.}

\item{target_meta}{Tibble.}

\item{vars_to_supp}{Character vector or \code{NULL}.}

\item{idvar}{Character or \code{NULL}.}

\item{qnam_rules}{Named list or \code{NULL}.}

\item{qlabel_source}{Character. Default \code{"metadata"}.}

\item{qorig_source}{Character. Default \code{"metadata"}.}

\item{qeval}{Character. Default \code{NA_character_}.}
}
\value{
Tibble or \code{NULL}.
}
\description{
Build SUPP-- dataset
}
