% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{build_ta_tv_te_ts_plugins}
\alias{build_ta_tv_te_ts_plugins}
\title{Trial design domain plugins}
\usage{
build_ta_tv_te_ts_plugins(
  domains = c("TA", "TV", "TE", "TS"),
  target_meta,
  config,
  rule_set,
  trial_design_data = list()
)
}
\arguments{
\item{domains}{Character vector.}

\item{target_meta}{Tibble.}

\item{config}{\code{sdtm_config}.}

\item{rule_set}{\code{rule_set}.}

\item{trial_design_data}{Named list.}
}
\value{
Named list of \code{build_result} objects.
}
\description{
Trial design domain plugins
}
