% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_rules.R
\name{canonicalize_rules}
\alias{canonicalize_rules}
\title{Canonicalize rules to standard form}
\usage{
canonicalize_rules(rule_set)
}
\arguments{
\item{rule_set}{\code{rule_set} object.}
}
\value{
Canonicalized \code{rule_set}.
}
\description{
Canonicalize rules to standard form
}
