% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g3_derive_dates.R
\name{combine_date_time}
\alias{combine_date_time}
\title{Combine date and time columns}
\usage{
combine_date_time(date, time, seconds = TRUE, tz = "UTC")
}
\arguments{
\item{date}{Parsed date tibble or character vector.}

\item{time}{Character vector.}

\item{seconds}{Logical. Default \code{TRUE}.}

\item{tz}{Character. Default \code{"UTC"}.}
}
\value{
Tibble.
}
\description{
Combine date and time columns
}
