% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_rules.R
\name{compile_rules}
\alias{compile_rules}
\title{Compile derivation rules from metadata}
\usage{
compile_rules(
  target_meta,
  source_meta = NULL,
  ct_lib = NULL,
  dsl = "json",
  strict = TRUE
)
}
\arguments{
\item{target_meta}{Tibble (validated, normalized).}

\item{source_meta}{Tibble or \code{NULL}. Optional source metadata.}

\item{ct_lib}{Tibble or \code{NULL}.}

\item{dsl}{Character. Default \code{"json"}.}

\item{strict}{Logical. Default \code{TRUE}.}
}
\value{
\code{rule_set} object.
}
\description{
Compile derivation rules from metadata
}
