% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_data_access.R
\name{convert_blanks_to_na}
\alias{convert_blanks_to_na}
\alias{convert_blanks_to_na.default}
\alias{convert_blanks_to_na.character}
\alias{convert_blanks_to_na.data.frame}
\title{Convert blanks to NA in a data frame}
\usage{
convert_blanks_to_na(x)

\method{convert_blanks_to_na}{default}(x)

\method{convert_blanks_to_na}{character}(x)

\method{convert_blanks_to_na}{data.frame}(x)
}
\arguments{
\item{x}{Any R object (method dispatch).}
}
\value{
Object of the same class with blanks converted to \code{NA}.
}
\description{
Turns empty strings (\code{""}) into proper R \code{NA} values. Preserves
all attributes (labels, etc.). This replicates the SAS concept of
"missing" for character variables.
}
\examples{
convert_blanks_to_na(c("hello", "", "world"))
# [1] "hello" NA      "world"

}
