% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g2_derive_ct.R
\name{decode_ct}
\alias{decode_ct}
\title{Decode controlled terminology}
\usage{
decode_ct(
  data,
  target_var,
  source_var,
  codelist_id,
  ct_lib,
  missing_decode = NA_character_
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{codelist_id}{Character.}

\item{ct_lib}{Tibble.}

\item{missing_decode}{Character. Default \code{NA_character_}.}
}
\value{
Tibble.
}
\description{
Decode controlled terminology
}
