% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{deduplicate_by_rule}
\alias{deduplicate_by_rule}
\title{Deduplicate records according to a rule}
\usage{
deduplicate_by_rule(
  data,
  keys,
  strategy = "first",
  order_by = NULL,
  priority_col = NULL,
  agg_fns = NULL
)
}
\arguments{
\item{data}{Tibble.}

\item{keys}{Character vector.}

\item{strategy}{Character. Default \code{"first"}.}

\item{order_by}{Character or \code{NULL}.}

\item{priority_col}{Character or \code{NULL}.}

\item{agg_fns}{Named list or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Deduplicate records according to a rule
}
