% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g6_derive_flags.R
\name{derive_baseline_flag}
\alias{derive_baseline_flag}
\title{Derive baseline flag}
\usage{
derive_baseline_flag(
  data,
  target_var = "BLFL",
  by = c("USUBJID"),
  baseline_visit = "BASELINE",
  visit_var = "VISIT"
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character. Default \code{"--BLFL"}.}

\item{by}{Character vector. Default \code{c("USUBJID","--TESTCD")}.}

\item{baseline_visit}{Character or numeric. Default \code{"BASELINE"}.}

\item{visit_var}{Character. Default \code{"VISIT"}.}
}
\value{
Tibble.
}
\description{
Derive baseline flag
}
