% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_case_when}
\alias{derive_case_when}
\title{Derive variable using multi-branch conditional logic}
\usage{
derive_case_when(data, target_var, conditions, default = NA)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{conditions}{Named list.}

\item{default}{Value. Default \code{NA}.}
}
\value{
Tibble.
}
\description{
Derive variable using multi-branch conditional logic
}
