% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_coalesce}
\alias{derive_coalesce}
\title{Derive variable as first non-missing across sources}
\usage{
derive_coalesce(data, target_var, sources, blank_to_na = TRUE)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{sources}{Character vector.}

\item{blank_to_na}{Logical. Default \code{TRUE}.}
}
\value{
Tibble.
}
\description{
Derive variable as first non-missing across sources
}
