% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_concat}
\alias{derive_concat}
\title{Concatenate multiple source columns}
\usage{
derive_concat(data, target_var, sources, sep = "", na_rm = TRUE, trim = TRUE)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{sources}{Character vector.}

\item{sep}{Character. Default \code{""}.}

\item{na_rm}{Logical. Default \code{TRUE}.}

\item{trim}{Logical. Default \code{TRUE}.}
}
\value{
Tibble.
}
\description{
Concatenate multiple source columns
}
