% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_constant}
\alias{derive_constant}
\title{Derive a constant value}
\usage{
derive_constant(data, target_var, value, when = NULL, type = NULL)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{value}{Scalar.}

\item{when}{Expression or \code{NULL}.}

\item{type}{Character or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Derive a constant value
}
