% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_data_access.R
\name{derive_core_keys}
\alias{derive_core_keys}
\title{Derive core identification keys (STUDYID, USUBJID)}
\usage{
derive_core_keys(data, config, subjid_col = "subjid", sep = "-")
}
\arguments{
\item{data}{Tibble.}

\item{config}{\code{sdtm_config}.}

\item{subjid_col}{Character. Default \code{"subjid"}.}

\item{sep}{Character. Default \code{"-"}.}
}
\value{
Tibble with STUDYID and USUBJID columns.
}
\description{
Derive core identification keys (STUDYID, USUBJID)
}
