% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{derive_domain_defaults}
\alias{derive_domain_defaults}
\title{Derive domain default variables}
\usage{
derive_domain_defaults(data, domain, config)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{config}{\code{sdtm_config}.}
}
\value{
Tibble.
}
\description{
Derive domain default variables
}
