% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g5_derive_ids.R
\name{derive_domain_keys}
\alias{derive_domain_keys}
\title{Derive STUDYID and DOMAIN constants}
\usage{
derive_domain_keys(data, domain, config, idvar = NULL, idvarval_col = NULL)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{config}{\code{sdtm_config}.}

\item{idvar}{Character or \code{NULL}.}

\item{idvarval_col}{Character or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Derive STUDYID and DOMAIN constants
}
