% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g3_derive_dates.R
\name{derive_duration}
\alias{derive_duration}
\title{Derive ISO 8601 duration}
\usage{
derive_duration(data, target_var, start_dtc, end_dtc, units = "auto")
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{start_dtc}{Character.}

\item{end_dtc}{Character.}

\item{units}{Character. Default \code{"auto"}.}
}
\value{
Tibble.
}
\description{
Derive ISO 8601 duration
}
