% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g3_derive_dates.R
\name{derive_dy}
\alias{derive_dy}
\title{Derive study day (--DY)}
\usage{
derive_dy(data, target_var, dtc_var, ref_var = "RFSTDTC", impute_policy = NULL)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{dtc_var}{Character.}

\item{ref_var}{Character. Default \code{"RFSTDTC"}.}

\item{impute_policy}{List or function or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Derive study day (--DY)
}
