% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g4_derive_visits.R
\name{derive_elapsed_time}
\alias{derive_elapsed_time}
\title{Derive elapsed time}
\usage{
derive_elapsed_time(data, target_var, start_var, end_var, units = "hours")
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{start_var}{Character.}

\item{end_var}{Character.}

\item{units}{Character. Default \code{"hours"}.}
}
\value{
Tibble.
}
\description{
Derive elapsed time
}
