% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g3_derive_dates.R
\name{derive_epoch}
\alias{derive_epoch}
\title{Derive EPOCH}
\usage{
derive_epoch(data, target_var = "EPOCH", dtc_var, epoch_map, by = "USUBJID")
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character. Default \code{"EPOCH"}.}

\item{dtc_var}{Character.}

\item{epoch_map}{Tibble or function.}

\item{by}{Character vector. Default \code{"USUBJID"}.}
}
\value{
Tibble.
}
\description{
Derive EPOCH
}
