% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g5_derive_ids.R
\name{derive_grpid}
\alias{derive_grpid}
\title{Derive Group Identifier}
\usage{
derive_grpid(data, target_var, group_by, method = "sequential")
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{group_by}{Character vector.}

\item{method}{Character. Default \code{"sequential"}.}
}
\value{
Tibble.
}
\description{
Derive Group Identifier
}
