% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_if_else}
\alias{derive_if_else}
\title{Derive variable using binary conditional logic}
\usage{
derive_if_else(
  data,
  target_var,
  condition,
  true_value,
  false_value,
  missing_value = NA
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{condition}{Expression or character.}

\item{true_value}{Value.}

\item{false_value}{Value.}

\item{missing_value}{Value. Default \code{NA}.}
}
\value{
Tibble.
}
\description{
Derive variable using binary conditional logic
}
