% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g6_derive_flags.R
\name{derive_lastobs_flag}
\alias{derive_lastobs_flag}
\title{Derive last observation flag}
\usage{
derive_lastobs_flag(data, target_var, by = "USUBJID", order_var)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{by}{Character vector. Default \code{"USUBJID"}.}

\item{order_var}{Character.}
}
\value{
Tibble.
}
\description{
Derive last observation flag
}
