% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_numeric_round}
\alias{derive_numeric_round}
\title{Round numeric values}
\usage{
derive_numeric_round(data, target_var, source_var, digits = 0L)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{digits}{Integer. Default \code{0L}.}
}
\value{
Tibble.
}
\description{
Round numeric values
}
