% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g6_derive_flags.R
\name{derive_occurrence}
\alias{derive_occurrence}
\title{Derive occurrence variable}
\usage{
derive_occurrence(
  data,
  target_var,
  source_var = NULL,
  present_value = "Y",
  absent_value = NA_character_
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{present_value}{Character. Default \code{"Y"}.}

\item{absent_value}{Character. Default \code{NA_character_}.}
}
\value{
Tibble.
}
\description{
Derive occurrence variable
}
