% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g6_derive_flags.R
\name{derive_reason}
\alias{derive_reason}
\title{Derive reason variable}
\usage{
derive_reason(data, target_var, source_var = NULL, stat_var = NULL)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{stat_var}{Character.}
}
\value{
Tibble.
}
\description{
Derive reason variable
}
