% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_regex_extract}
\alias{derive_regex_extract}
\title{Extract values using regex}
\usage{
derive_regex_extract(
  data,
  target_var,
  source_var,
  pattern,
  group = 1L,
  transform = NULL
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{pattern}{Character.}

\item{group}{Integer. Default \code{1L}.}

\item{transform}{Function or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Extract values using regex
}
