% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_regex_replace}
\alias{derive_regex_replace}
\title{Replace text patterns using regex}
\usage{
derive_regex_replace(
  data,
  target_var,
  source_var,
  pattern,
  replacement,
  all = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{pattern}{Character.}

\item{replacement}{Character.}

\item{all}{Logical. Default \code{TRUE}.}
}
\value{
Tibble.
}
\description{
Replace text patterns using regex
}
