% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g4_derive_visits.R
\name{derive_rel_time}
\alias{derive_rel_time}
\title{Derive relative time}
\usage{
derive_rel_time(data, target_var, dtc_var, ref_dtc_var)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{dtc_var}{Character.}

\item{ref_dtc_var}{Character.}
}
\value{
Tibble.
}
\description{
Derive relative time
}
