% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g5_derive_ids.R
\name{derive_seq}
\alias{derive_seq}
\title{Derive sequence number (--SEQ)}
\usage{
derive_seq(
  data,
  target_var,
  by = "USUBJID",
  order_by,
  ties = "dense",
  start_at = 1L
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{by}{Character vector. Default \code{"USUBJID"}.}

\item{order_by}{Character vector.}

\item{ties}{Character. Default \code{"dense"}.}

\item{start_at}{Integer. Default \code{1L}.}
}
\value{
Tibble.
}
\description{
Derive sequence number (--SEQ)
}
