% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g5_derive_ids.R
\name{derive_spid}
\alias{derive_spid}
\title{Derive Sponsor-Defined Identifier}
\usage{
derive_spid(data, target_var, source_cols, sep = "-", prefix = NULL)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_cols}{Character vector.}

\item{sep}{Character. Default \code{"-"}.}

\item{prefix}{Character or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Derive Sponsor-Defined Identifier
}
