% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g6_derive_flags.R
\name{derive_status}
\alias{derive_status}
\title{Derive status variable}
\usage{
derive_status(
  data,
  target_var,
  result_var,
  done_value = NA_character_,
  not_done_value = "NOT DONE"
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{result_var}{Character.}

\item{done_value}{Character. Default \code{NA_character_}.}

\item{not_done_value}{Character. Default \code{"NOT DONE"}.}
}
\value{
Tibble.
}
\description{
Derive status variable
}
