% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g4_derive_visits.R
\name{derive_tpt}
\alias{derive_tpt}
\title{Derive timepoint}
\usage{
derive_tpt(data, target_var, source_var = NULL, tpt_map = NULL)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character. Default \code{"--TPT"}.}

\item{source_var}{Character.}

\item{tpt_map}{Named character vector.}
}
\value{
Tibble.
}
\description{
Derive timepoint
}
