% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_trim_pad}
\alias{derive_trim_pad}
\title{Trim or pad a string variable}
\usage{
derive_trim_pad(
  data,
  target_var,
  source_var,
  width = NULL,
  side = "both",
  pad = " "
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{width}{Integer or \code{NULL}.}

\item{side}{Character. Default \code{"both"}.}

\item{pad}{Character. Default \code{" "}.}
}
\value{
Tibble.
}
\description{
Trim or pad a string variable
}
