% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{derive_unit_standardize}
\alias{derive_unit_standardize}
\title{Standardize units}
\usage{
derive_unit_standardize(data, target_var, source_var, unit_map = list())
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{unit_map}{Named list.}
}
\value{
Tibble.
}
\description{
Standardize units
}
