% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g5_derive_ids.R
\name{derive_usubjid}
\alias{derive_usubjid}
\title{Derive USUBJID}
\usage{
derive_usubjid(
  data,
  studyid,
  subjid_col = "subjid",
  sep = "-",
  validate_existing = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{studyid}{Character.}

\item{subjid_col}{Character. Default \code{"subjid"}.}

\item{sep}{Character. Default \code{"-"}.}

\item{validate_existing}{Logical. Default \code{TRUE}.}
}
\value{
Tibble.
}
\description{
Derive USUBJID
}
