% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{derive_variable}
\alias{derive_variable}
\title{Single variable derivation dispatcher}
\usage{
derive_variable(data, var, rule, context)
}
\arguments{
\item{data}{Tibble.}

\item{var}{Character.}

\item{rule}{List.}

\item{context}{List or \code{build_context}.}
}
\value{
Tibble.
}
\description{
Single variable derivation dispatcher
}
