% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g4_derive_visits.R
\name{derive_visit}
\alias{derive_visit}
\title{Derive VISIT from visit map}
\usage{
derive_visit(
  data,
  target_var = "VISIT",
  visit_map,
  dy_var = NULL,
  by = "USUBJID"
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character. Default \code{"VISIT"}.}

\item{visit_map}{Tibble.}

\item{dy_var}{Character. Study day variable.}

\item{by}{Character vector. Default \code{"USUBJID"}.}
}
\value{
Tibble.
}
\description{
Derive VISIT from visit map
}
