% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g4_derive_visits.R
\name{derive_visitdy}
\alias{derive_visitdy}
\title{Derive VISITDY}
\usage{
derive_visitdy(
  data,
  target_var = "VISITDY",
  visit_var = "VISIT",
  dy_var = NULL,
  visit_map = NULL
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character. Default \code{"VISITDY"}.}

\item{visit_var}{Character. Default \code{"VISIT"}.}

\item{dy_var}{Character. Study day variable.}
}
\value{
Tibble.
}
\description{
Derive VISITDY
}
