% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g4_derive_visits.R
\name{derive_visitnum}
\alias{derive_visitnum}
\title{Derive VISITNUM}
\usage{
derive_visitnum(data, target_var = "VISITNUM", visit_map, visit_var = "VISIT")
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character. Default \code{"VISITNUM"}.}

\item{visit_map}{Tibble.}

\item{visit_var}{Character. Default \code{"VISIT"}.}
}
\value{
Tibble.
}
\description{
Derive VISITNUM
}
