% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_dependency.R
\name{detect_cycles}
\alias{detect_cycles}
\title{Detect cycles in a dependency graph}
\usage{
detect_cycles(graph)
}
\arguments{
\item{graph}{\code{igraph} directed graph.}
}
\value{
List of cycle paths.
}
\description{
Detect cycles in a dependency graph
}
