% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{.build_derivation_chain}
\alias{.build_derivation_chain}
\title{Build the piped derivation chain}
\usage{
.build_derivation_chain(var_order, dom_rules, domain, config, include_comments)
}
\description{
Build the piped derivation chain
}
\keyword{internal}
