% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{.categorize_var}
\alias{.categorize_var}
\title{Assign a semantic category to a variable based on name and rule type}
\usage{
.categorize_var(var, rule_type, domain)
}
\description{
Assign a semantic category to a variable based on name and rule type
}
\keyword{internal}
