% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{.detect_source_datasets}
\alias{.detect_source_datasets}
\title{Detect unique source datasets referenced by domain rules}
\usage{
.detect_source_datasets(dom_rules)
}
\description{
Detect unique source datasets referenced by domain rules
}
\keyword{internal}
