% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{.is_pipe_step}
\alias{.is_pipe_step}
\title{Test whether a rule type requires its own pipe step}
\usage{
.is_pipe_step(rule_type)
}
\description{
Test whether a rule type requires its own pipe step
}
\keyword{internal}
