% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{.render_mutate_expr}
\alias{.render_mutate_expr}
\title{Render a mutate-compatible expression for a rule}
\usage{
.render_mutate_expr(var, rule, config)
}
\description{
Render a mutate-compatible expression for a rule
}
\keyword{internal}
