% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{.render_pipe_step}
\alias{.render_pipe_step}
\title{Render a pipe-step function call for a rule}
\usage{
.render_pipe_step(var, rule, config)
}
\description{
Render a pipe-step function call for a rule
}
\keyword{internal}
