% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_rules.R
\name{enrich_rules_with_ct}
\alias{enrich_rules_with_ct}
\title{Enrich rules with CT details}
\usage{
enrich_rules_with_ct(rule_set, ct_lib)
}
\arguments{
\item{rule_set}{\code{rule_set} object.}

\item{ct_lib}{Tibble.}
}
\value{
Updated \code{rule_set}.
}
\description{
Enrich rules with CT details
}
