% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{expand_checkbox}
\alias{expand_checkbox}
\title{Expand checkbox columns into records}
\usage{
expand_checkbox(
  data,
  cols,
  map,
  id_cols,
  checked_values = c("Y", "1", "TRUE", "Yes", "X"),
  drop_unchecked = TRUE,
  unchecked_flag = NULL
)
}
\arguments{
\item{data}{Tibble.}

\item{cols}{Character vector.}

\item{map}{Named list.}

\item{id_cols}{Character vector.}

\item{checked_values}{Character vector.}

\item{drop_unchecked}{Logical. Default \code{TRUE}.}

\item{unchecked_flag}{Character or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Expand checkbox columns into records
}
