% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{expand_visits}
\alias{expand_visits}
\title{Expand records across visit schedule}
\usage{
expand_visits(data, visit_map, by, filter_fn = NULL)
}
\arguments{
\item{data}{Tibble.}

\item{visit_map}{Tibble.}

\item{by}{Character vector.}

\item{filter_fn}{Function or \code{NULL}.}
}
\value{
Tibble.
}
\description{
Expand records across visit schedule
}
