% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_k_export.R
\name{export_rds_csv}
\alias{export_rds_csv}
\title{Export a domain to RDS and CSV}
\usage{
export_rds_csv(data, domain, output_dir, formats = c("rds", "csv"))
}
\arguments{
\item{data}{Tibble. Domain dataset.}

\item{domain}{Character.}

\item{output_dir}{Character.}

\item{formats}{Character vector. Subset of \code{c("rds","csv","rda")}.}
}
\value{
Invisible named list of paths.
}
\description{
Export a domain to RDS and CSV
}
