% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g3_derive_dates.R
\name{format_iso_dtc}
\alias{format_iso_dtc}
\title{Format parsed date/time as ISO 8601 --DTC string}
\usage{
format_iso_dtc(parsed, time = NULL, keep_partial = TRUE, impute = NULL)
}
\arguments{
\item{parsed}{Tibble from parse_partial_date or combine_date_time,
OR a character vector of raw date strings.}

\item{time}{Character vector or \code{NULL}.}

\item{keep_partial}{Logical. Default \code{TRUE}.}

\item{impute}{Function or \code{NULL}.}
}
\value{
Character vector of ISO 8601 strings.
}
\description{
Format parsed date/time as ISO 8601 --DTC string
}
