% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{gen_project_scaffold}
\alias{gen_project_scaffold}
\title{Generate project scaffold (directories and template files)}
\usage{
gen_project_scaffold(output_dir, config, domains)
}
\arguments{
\item{output_dir}{Character.}

\item{config}{\code{sdtm_config}.}

\item{domains}{Character vector.}
}
\value{
Invisible \code{NULL}.
}
\description{
Generate project scaffold (directories and template files)
}
