% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{gen_qmd_domain}
\alias{gen_qmd_domain}
\title{Generate a Quarto domain report}
\usage{
gen_qmd_domain(
  domain,
  rule_set,
  target_meta,
  build_result = NULL,
  output_path = NULL
)
}
\arguments{
\item{domain}{Character.}

\item{rule_set}{\code{rule_set}.}

\item{target_meta}{Tibble.}

\item{build_result}{Named list.}

\item{output_path}{Character or \code{NULL}.}
}
\value{
Character.
}
\description{
Generate a Quarto domain report
}
