% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{gen_shared_utils_script}
\alias{gen_shared_utils_script}
\title{Generate shared utility script}
\usage{
gen_shared_utils_script(config, output_path = NULL)
}
\arguments{
\item{config}{\code{sdtm_config}.}

\item{output_path}{Character or \code{NULL}.}
}
\value{
Character.
}
\description{
Generate shared utility script
}
