% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_data_access.R
\name{get_subject_level}
\alias{get_subject_level}
\title{Get subject-level data}
\usage{
get_subject_level(
  raw_data,
  config,
  dm_dataset = "dm",
  key_vars = c("USUBJID", "RFSTDTC", "RFENDTC", "ARM", "ACTARM")
)
}
\arguments{
\item{raw_data}{Named list of tibbles.}

\item{config}{\code{sdtm_config}.}

\item{dm_dataset}{Character. Default \code{"dm"}.}

\item{key_vars}{Character vector.}
}
\value{
Tibble: one row per subject.
}
\description{
Get subject-level data
}
