% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_study.R
\name{get_template_config}
\alias{get_template_config}
\title{Get the template config.yaml path}
\usage{
get_template_config(copy_to = NULL)
}
\arguments{
\item{copy_to}{Character or \code{NULL}. If provided, copies the template to
this path. The directory will be created if it doesn't exist.}
}
\value{
Invisible path to the template file (or the copy destination).
}
\description{
Returns the path to the built-in template \code{config.yaml} that can be copied
to your study folder as a starting point.
}
\examples{
# Just get the path
get_template_config()

# Copy to your study folder
get_template_config(copy_to = "my_study/metadata/config.yaml")

}
