% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_rules.R
\name{infer_rule_dependencies}
\alias{infer_rule_dependencies}
\title{Infer inter-variable dependencies from rules}
\usage{
infer_rule_dependencies(rule_set)
}
\arguments{
\item{rule_set}{\code{rule_set} object.}
}
\value{
Updated \code{rule_set}.
}
\description{
Infer inter-variable dependencies from rules
}
