% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_data_access.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load raw datasets from files}
\usage{
load_raw_data(
  paths,
  formats = NULL,
  encoding = "UTF-8",
  db_con = NULL,
  db_tables = NULL
)
}
\arguments{
\item{paths}{Named character vector.}

\item{formats}{Named character vector or \code{NULL}.}

\item{encoding}{Character. Default \code{"UTF-8"}.}

\item{db_con}{DBI connection or \code{NULL}.}

\item{db_tables}{Named character vector.}
}
\value{
Named list of tibbles.
}
\description{
Load raw datasets from files
}
