% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g1_derive_map.R
\name{map_direct}
\alias{map_direct}
\title{Direct mapping from source column}
\usage{
map_direct(
  data,
  target_var,
  source_var,
  transform = NULL,
  type = NULL,
  blank_to_na = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{transform}{Function or \code{NULL}.}

\item{type}{Character or \code{NULL}.}

\item{blank_to_na}{Logical. Default \code{TRUE}.}
}
\value{
Tibble.
}
\description{
Direct mapping from source column
}
