% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g2_derive_ct.R
\name{map_unknown}
\alias{map_unknown}
\title{Map unknown indicators to SDTM conventions}
\usage{
map_unknown(
  data,
  target_var,
  source_var,
  unknown_tokens = c("UNK", "UNKNOWN", "NK", "U", "N/A"),
  sdtm_value = "UNKNOWN"
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{unknown_tokens}{Character vector.}

\item{sdtm_value}{Character. Default \code{"UNKNOWN"}.}
}
\value{
Tibble.
}
\description{
Map unknown indicators to SDTM conventions
}
