% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g2_derive_ct.R
\name{map_yes_no}
\alias{map_yes_no}
\title{Map values to Y/N format}
\usage{
map_yes_no(
  data,
  target_var,
  source_var,
  yes_values = c("Y", "YES", "1", "TRUE", "X", "CHECKED"),
  no_values = c("N", "NO", "0", "FALSE", "UNCHECKED"),
  na_policy = "na"
)
}
\arguments{
\item{data}{Tibble.}

\item{target_var}{Character.}

\item{source_var}{Character.}

\item{yes_values}{Character vector.}

\item{no_values}{Character vector.}

\item{na_policy}{Character. Default \code{"na"}.}
}
\value{
Tibble.
}
\description{
Map values to Y/N format
}
