% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_rules.R
\name{parse_rule_dsl}
\alias{parse_rule_dsl}
\title{Parse a rule DSL specification}
\usage{
parse_rule_dsl(rule_text, var = NA_character_)
}
\arguments{
\item{rule_text}{Character.}

\item{var}{Character. Target variable name.}
}
\value{
Named list.
}
\description{
Parse a rule DSL specification
}
