% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_rules.R
\name{parse_rule_json}
\alias{parse_rule_json}
\title{Parse a rule JSON specification}
\usage{
parse_rule_json(rule_json, var = NA_character_)
}
\arguments{
\item{rule_json}{Character or list.}

\item{var}{Character. Target variable name.}
}
\value{
Named list.
}
\description{
Parse a rule JSON specification
}
