% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_dependency.R
\name{plan_build_steps}
\alias{plan_build_steps}
\title{Generate a domain-level build plan}
\usage{
plan_build_steps(domain, dep_graph, rule_set)
}
\arguments{
\item{domain}{Character.}

\item{dep_graph}{Output of \code{build_dependency_graph}.}

\item{rule_set}{\code{rule_set}.}
}
\value{
Tibble with build steps.
}
\description{
Generate a domain-level build plan
}
