% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{render_rule_code}
\alias{render_rule_code}
\title{Render R code for a single rule}
\usage{
render_rule_code(var, rule, style = "tidyverse")
}
\arguments{
\item{var}{Character. Variable name.}

\item{rule}{List. Rule object.}

\item{style}{Character.}
}
\value{
Character.
}
\description{
Render R code for a single rule
}
\keyword{internal}
