% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{render_rule_comments}
\alias{render_rule_comments}
\title{Render rule as code comment}
\usage{
render_rule_comments(rule)
}
\arguments{
\item{rule}{List.}
}
\value{
Character.
}
\description{
Render rule as code comment
}
