% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_b_metadata.R
\name{resolve_domain_model}
\alias{resolve_domain_model}
\title{Resolve domain model from metadata}
\usage{
resolve_domain_model(domain, target_meta, config, ig_version = "3.4")
}
\arguments{
\item{domain}{Character. Domain abbreviation (e.g., \code{"AE"}).}

\item{target_meta}{Tibble. Full target metadata.}

\item{config}{\code{sdtm_config}.}

\item{ig_version}{Character. SDTM IG version. Default \code{"3.4"}.}
}
\value{
Tibble filtered and ordered for the domain.
}
\description{
Extracts and optionally sorts the metadata rows for a single domain.
}
