% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{resolve_keys}
\alias{resolve_keys}
\title{Resolve domain keys}
\usage{
resolve_keys(data, domain, target_meta, config)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{target_meta}{Tibble.}

\item{config}{\code{sdtm_config}.}
}
\value{
Named list.
}
\description{
Resolve domain keys
}
