% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_dependency.R
\name{resolve_late_binding}
\alias{resolve_late_binding}
\title{Resolve late-binding variables}
\usage{
resolve_late_binding(graph, cycles, late_bind_vars)
}
\arguments{
\item{graph}{\code{igraph} graph.}

\item{cycles}{List of cycle paths.}

\item{late_bind_vars}{Character vector.}
}
\value{
Named list.
}
\description{
Resolve late-binding variables
}
