% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{safe_join}
\alias{safe_join}
\title{Safe join with cardinality assertions}
\usage{
safe_join(
  x,
  y,
  by,
  type = "left",
  cardinality = "1:1",
  on_violation = "error",
  dedupe_rule = NULL,
  merge_indicator = NULL,
  suffix = c(".x", ".y")
)
}
\arguments{
\item{x}{Tibble. Left dataset.}

\item{y}{Tibble. Right dataset.}

\item{by}{Character vector. Join keys.}

\item{type}{Character. Default \code{"left"}.}

\item{cardinality}{Character. Default \code{"1:1"}.}

\item{on_violation}{Character. Default \code{"error"}.}

\item{dedupe_rule}{List or \code{NULL}.}

\item{merge_indicator}{Character or \code{NULL}.}

\item{suffix}{Character vector. Default \code{c(".x", ".y")}.}
}
\value{
Tibble.
}
\description{
Safe join with cardinality assertions
}
