% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_j_codegen.R
\name{serialize_rules_to_yaml}
\alias{serialize_rules_to_yaml}
\title{Serialize rules to YAML}
\usage{
serialize_rules_to_yaml(rule_set, output_path = NULL)
}
\arguments{
\item{rule_set}{\code{rule_set}.}

\item{output_path}{Character or \code{NULL}.}
}
\value{
Character.
}
\description{
Serialize rules to YAML
}
