% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_f_joins.R
\name{split_records}
\alias{split_records}
\title{Split delimited values into multiple records}
\usage{
split_records(
  data,
  col,
  sep = ";",
  trim = TRUE,
  id_cols = NULL,
  new_col = NULL,
  drop_empty = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{col}{Character.}

\item{sep}{Character. Default \code{";"}.}

\item{trim}{Logical. Default \code{TRUE}.}

\item{id_cols}{Character vector or \code{NULL}.}

\item{new_col}{Character or \code{NULL}.}

\item{drop_empty}{Logical. Default \code{TRUE}.}
}
\value{
Tibble.
}
\description{
Split delimited values into multiple records
}
