% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_data_access.R
\name{standardize_types}
\alias{standardize_types}
\title{Coerce column types}
\usage{
standardize_types(
  data,
  dataset_name = NA_character_,
  coerce_factors = TRUE,
  strip_labels = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{dataset_name}{Character.}

\item{coerce_factors}{Logical. Default \code{TRUE}.}

\item{strip_labels}{Logical. Default \code{TRUE}.}
}
\value{
Tibble with corrected types.
}
\description{
Converts factors to character and strips haven labels.
}
