% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{summarize_validation_report}
\alias{summarize_validation_report}
\title{Summary of a validation report}
\usage{
summarize_validation_report(report, verbose = TRUE)
}
\arguments{
\item{report}{\code{validation_report}.}

\item{verbose}{Logical. Default \code{TRUE}.}
}
\value{
Tibble summary.
}
\description{
Summary of a validation report
}
