% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_dependency.R
\name{topo_sort_rules}
\alias{topo_sort_rules}
\title{Topologically sort rules within a domain}
\usage{
topo_sort_rules(graph)
}
\arguments{
\item{graph}{An \code{igraph} directed graph or output of \code{build_dependency_graph}.}
}
\value{
Character vector of variable names in derivation order.
}
\description{
Topologically sort rules within a domain
}
