% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_cross_domain}
\alias{validate_cross_domain}
\title{Cross-domain validation}
\usage{
validate_cross_domain(domain_data, config)
}
\arguments{
\item{domain_data}{Named list of domain tibbles.}

\item{config}{\code{sdtm_config}.}
}
\value{
\code{validation_report}.
}
\description{
Cross-domain validation
}
