% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_ct_conformance}
\alias{validate_ct_conformance}
\title{Validate controlled terminology conformance}
\usage{
validate_ct_conformance(data, target_meta, domain, ct_lib = NULL, report)
}
\arguments{
\item{data}{Tibble.}

\item{target_meta}{Tibble.}

\item{domain}{Character.}

\item{ct_lib}{Tibble or \code{NULL}.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate controlled terminology conformance
}
