% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_b_metadata.R
\name{validate_ct_library}
\alias{validate_ct_library}
\title{Validate controlled terminology library}
\usage{
validate_ct_library(ct_lib)
}
\arguments{
\item{ct_lib}{Tibble. CT library to validate.}
}
\value{
Invisible \code{ct_lib} (for piping).
}
\description{
Checks that required columns (\code{codelist_id}, \code{coded_value}) exist.
}
