% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_g2_derive_ct.R
\name{validate_ct_values}
\alias{validate_ct_values}
\title{Validate CT values}
\usage{
validate_ct_values(
  data,
  var,
  codelist_id,
  ct_lib,
  check_type = "coded",
  allow_na = TRUE,
  allow_other = character()
)
}
\arguments{
\item{data}{Tibble.}

\item{var}{Character.}

\item{codelist_id}{Character.}

\item{ct_lib}{Tibble.}

\item{check_type}{Character. Default \code{"coded"}.}

\item{allow_na}{Logical. Default \code{TRUE}.}

\item{allow_other}{Character vector.}
}
\value{
Tibble of findings.
}
\description{
Validate CT values
}
