% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_domain_value}
\alias{validate_domain_value}
\title{Validate DOMAIN variable value matches domain abbreviation}
\usage{
validate_domain_value(data, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate DOMAIN variable value matches domain abbreviation
}
