% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_iso8601}
\alias{validate_iso8601}
\title{Validate ISO 8601 date/time variables}
\usage{
validate_iso8601(data, target_meta, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{target_meta}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate ISO 8601 date/time variables
}
