% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_keys_unique}
\alias{validate_keys_unique}
\title{Validate key variables are unique}
\usage{
validate_keys_unique(data, target_meta, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{target_meta}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate key variables are unique
}
