% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_lengths_types_labels}
\alias{validate_lengths_types_labels}
\title{Validate lengths, types, and labels}
\usage{
validate_lengths_types_labels(data, target_meta, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{target_meta}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate lengths, types, and labels
}
