% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_no_allna_reqexp}
\alias{validate_no_allna_reqexp}
\title{Validate no Required/Expected variable is entirely NA}
\usage{
validate_no_allna_reqexp(data, target_meta, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{target_meta}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate no Required/Expected variable is entirely NA
}
