% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_no_duplicate_rows}
\alias{validate_no_duplicate_rows}
\title{Validate no fully duplicate rows}
\usage{
validate_no_duplicate_rows(data, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate no fully duplicate rows
}
