% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_required_vars}
\alias{validate_required_vars}
\title{Validate required variables are present and non-missing}
\usage{
validate_required_vars(data, target_meta, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{target_meta}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate required variables are present and non-missing
}
