% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_rules.R
\name{validate_rules}
\alias{validate_rules}
\title{Validate compiled rules}
\usage{
validate_rules(rule_set, source_meta, ct_lib = NULL)
}
\arguments{
\item{rule_set}{\code{rule_set} object.}

\item{source_meta}{Tibble.}

\item{ct_lib}{Tibble or \code{NULL}.}
}
\value{
Updated \code{rule_set}.
}
\description{
Validate compiled rules
}
