% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_seq_integrity}
\alias{validate_seq_integrity}
\title{Validate --SEQ is a positive integer with no gaps per subject}
\usage{
validate_seq_integrity(data, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate --SEQ is a positive integer with no gaps per subject
}
