% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_studyid_constant}
\alias{validate_studyid_constant}
\title{Validate STUDYID is constant across all rows}
\usage{
validate_studyid_constant(data, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate STUDYID is constant across all rows
}
