% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_i_validation.R
\name{validate_value_level}
\alias{validate_value_level}
\title{Validate value-level metadata}
\usage{
validate_value_level(data, vlm_meta = NULL, domain, report)
}
\arguments{
\item{data}{Tibble.}

\item{vlm_meta}{Tibble or \code{NULL}.}

\item{domain}{Character.}

\item{report}{\code{validation_report}.}
}
\value{
Updated \code{validation_report}.
}
\description{
Validate value-level metadata
}
