% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_k_export.R
\name{write_codelist_support}
\alias{write_codelist_support}
\title{Write codelist support file}
\usage{
write_codelist_support(ct_lib, output_path)
}
\arguments{
\item{ct_lib}{Tibble.}

\item{output_path}{Character.}
}
\value{
Invisible file path.
}
\description{
Write codelist support file
}
