% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_k_export.R
\name{write_define_support}
\alias{write_define_support}
\title{Write define.xml support file}
\usage{
write_define_support(domains, target_meta, config, output_path)
}
\arguments{
\item{domains}{Named list of built domain tibbles.}

\item{target_meta}{Tibble.}

\item{config}{\code{sdtm_config}.}

\item{output_path}{Character.}
}
\value{
Invisible file path.
}
\description{
Generates a CSV that maps variables to metadata for define.xml generation.
}
