% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_k_export.R
\name{write_origin_support}
\alias{write_origin_support}
\title{Write origin metadata support file}
\usage{
write_origin_support(target_meta, output_path)
}
\arguments{
\item{target_meta}{Tibble.}

\item{output_path}{Character.}
}
\value{
Invisible file path.
}
\description{
Write origin metadata support file
}
