% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_k_export.R
\name{write_value_level_support}
\alias{write_value_level_support}
\title{Write value-level metadata support file}
\usage{
write_value_level_support(vlm = NULL, output_path)
}
\arguments{
\item{vlm}{Tibble or \code{NULL}.}

\item{output_path}{Character.}
}
\value{
Invisible file path or \code{NULL}.
}
\description{
Write value-level metadata support file
}
