% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_h_builders.R
\name{finalize_domain}
\alias{finalize_domain}
\title{Finalize a domain dataset for export}
\usage{
finalize_domain(
  data,
  domain,
  target_meta,
  config,
  create_seq = TRUE,
  create_supp = TRUE,
  strict_labels = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{domain}{Character.}

\item{target_meta}{Tibble.}

\item{config}{\code{sdtm_config}.}

\item{create_seq}{Logical. Default \code{TRUE}.}

\item{create_supp}{Logical. Default \code{TRUE}.}

\item{strict_labels}{Logical. Default \code{TRUE}.}
}
\value{
Named list.
}
\description{
Finalize a domain dataset for export
}
