% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_data_access.R
\name{standardize_names}
\alias{standardize_names}
\title{Standardize column names}
\usage{
standardize_names(
  data,
  dataset_name = NA_character_,
  to_lower = TRUE,
  replace_special = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{dataset_name}{Character.}

\item{to_lower}{Logical. Default \code{TRUE}.}

\item{replace_special}{Logical. Default \code{TRUE}.}
}
\value{
Tibble with standardized names.
}
\description{
Standardize column names
}
